msm: clock-mdss: Do not toggle lock detect during handoff
If continuous splash screen feature is enabled, then the DSI
clocks are left on in the bootloader. When the kernel boots
up, te DSI PLL lock status is checked in order to determine
whether these clocks should be handed off or not. Ensure that
the lock detect register is not toggled at this time to avoid
any blinks in the display.

Change-Id: Ica8c42e260d9b3ceffda885b1942fab945f60c86
CRs-Fixed: 506037
Signed-off-by: Aravind Venkateswaran <aravindh@codeaurora.org>