USB KEYBOARD/MOUSE: Added on/off trigger to driver. READ MORE.
When was ON by default I have seen that USB is always HID device.

And i have always received 2 connection sounds and vibrate...

It's annoying! so now all good, one alert for connection, and USB
is Android Device, and when I have connect USB Keyboard or mouse or both
via wireless connector (Microsoft keyboard + mouse Infra-Red)

it's works! phone detect mouse + keyboard and i can use them on screen!

the On/off trigger seems not needed to use the HID and can be OFF.

I have left it for future use, for compatibility if needed.